updated controller routes

+++ Controllers/DiscordGuildController.cs
@@ -17,16 +17,26 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/characters/{guildId}")]
-        public ActionResult GetCharactersByGuildId(long guildId = 1077311704985239700)
+        public ActionResult GetCharactersByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Character> characters = _repository.GetCharactersByGuildId(guildId);
 
             return Ok(characters);
         }
 
         [HttpGet("/guilds/character-races/{guildId}")]
-        public ActionResult<IList<CharacterRace>> GetCharacterRacesByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetCharacterRacesByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<CharacterRace> characterRaces = _repository.GetCharacterRacesByGuildId(guildId);
             IList<CharacterSheetPropertyDto> characterRaceDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterRaces);
 
@@ -34,8 +44,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/character-classes/{guildId}")]
-        public ActionResult<IList<CharacterClass>> GetClassesByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetClassesByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<CharacterClass> characterClasses = _repository.GetCharacterClassesByGuildId(guildId);
             IList<CharacterSheetPropertyDto> characterClassDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -43,8 +58,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/features/{guildId}")]
-        public ActionResult<IList<Feature>> GetFeaturesByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetFeaturesByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Feature> characterClasses = _repository.GetFeaturesByGuildId(guildId);
             IList<CharacterSheetPropertyDto> featureDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -52,8 +72,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/traits/{guildId}")]
-        public ActionResult<IList<Trait>> GetTraitsByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetTraitsByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Trait> characterClasses = _repository.GetTraitsByGuildId(guildId);
             IList<CharacterSheetPropertyDto> traitDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -61,8 +86,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/organizations/{guildId}")]
-        public ActionResult<IList<Organization>> GetOrganizationsByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetOrganizationsByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Organization> characterClasses = _repository.GetOrganizationsByGuildId(guildId);
             IList<CharacterSheetPropertyDto> organizationDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -70,8 +100,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/items/{guildId}")]
-        public ActionResult<IList<Item>> GetItemsByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetItemsByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Item> characterClasses = _repository.GetItemsByGuildId(guildId);
             IList<CharacterSheetPropertyDto> itemDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -79,8 +114,13 @@ namespace DMApp.Data
         }
 
         [HttpGet("/guilds/spells/{guildId}")]
-        public ActionResult<IList<Spell>> GetspellsByGuildId(long guildId = 1077311704985239684)
+        public ActionResult<IList<CharacterSheetPropertyDto>> GetspellsByGuildId(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             IList<Spell> characterClasses = _repository.GetSpellsByGuildId(guildId);
             IList<CharacterSheetPropertyDto> spellDtos = _mapper.Map<IList<CharacterSheetPropertyDto>>(characterClasses);
 
@@ -88,24 +128,28 @@ namespace DMApp.Data
         }
 
         [HttpPost("/guilds/{guildId}")]
-        public ActionResult AddGuild(long guildId = 1077311704985239684)
+        public ActionResult AddGuild(long guildId = 0)
         {
+            if (guildId == 0)
+            {
+                long.TryParse(Environment.GetEnvironmentVariable("DefaultGuildId"), out guildId);
+            }
+
             DiscordGuild guild = _repository.CreateGuild(guildId);
             if (_repository.SaveChanges()) return Ok(guild);
             else return BadRequest(new { error = $"Unable to save guild: {guildId}" });
         }
 
         [HttpDelete("/guilds/{guildId}")]
-        public ActionResult DeleteGuild(long guildId = 1077311704985239700)
+        public ActionResult DeleteGuild(long guildId)
         {
+            if(guildId == null || guildId == 0) return BadRequest(new { error = $"Unable to delete guild: {guildId}" });
+
             _repository.DeleteGuild(guildId);
 
-            if (_repository.SaveChanges()) return Ok(new {
-                    status = 200,
-                    message = $"Guild successfully deleted: {guildId}"
-                }
-            );
-            else return BadRequest(new { error = $"Unable to delete guild: {guildId}" });
+            if (_repository.SaveChanges()) return Ok(new { status = 200,Message = $"Guild successfully deleted: {guildId}" });
+
+            return BadRequest(new { error = $"Unable to delete guild: {guildId}" });
         }
     }
 }
